split geometry calculation into multies files
