+++ src/CMakeLists.txt
@@ -89,7 +89,9 @@ if(NeoN_WITH_VTK)
             "mesh/unstructured/io/connectivity/cellReconstruction.cpp"
             "mesh/unstructured/io/connectivity/nodeOrdering.cpp"
             "mesh/unstructured/io/connectivity/vtkExtraction.cpp"
-            "mesh/unstructured/io/meshGeometry.cpp"
+            "mesh/unstructured/io/geometry/faceGeometry.cpp"
+            "mesh/unstructured/io/geometry/cellGeometry.cpp"
+            "mesh/unstructured/io/geometry/computeGeometry.cpp"
             "mesh/unstructured/io/meshConverter.cpp"
             "mesh/unstructured/io/cgnsMeshReader.cpp"
             "mesh/unstructured/io/cgnsMeshWriter.cpp"
