+++ src/mesh/unstructured/io/geometry/cellGeometry.cpp
@@ -2,13 +2,12 @@
 //
 // SPDX-License-Identifier: MIT
 
-#include "NeoN/mesh/unstructured/io/meshGeometry.hpp"
+#include "NeoN/mesh/unstructured/io/geometry/cellGeometry.hpp"
 
 #include "NeoN/core/containerFreeFunctions.hpp"
 #include "NeoN/core/parallelAlgorithms.hpp"
 
 #include <cmath>
-#include <vector>
 
 
 namespace NeoN::io
@@ -16,18 +15,6 @@ namespace NeoN::io
 
 namespace
 {
-
-NEON_INLINE_FUNCTION
-Vec3 triangleArea(const Vec3& p0, const Vec3& p1, const Vec3& p2)
-{
-    Vec3 e1 = p1 - p0;
-    Vec3 e2 = p2 - p0;
-    return Vec3 {
-        0.5 * (e1[1] * e2[2] - e1[2] * e2[1]),
-        0.5 * (e1[2] * e2[0] - e1[0] * e2[2]),
-        0.5 * (e1[0] * e2[1] - e1[1] * e2[0])
-    };
-}
 
 NEON_INLINE_FUNCTION
 scalar tetVolume(const Vec3& p0, const Vec3& p1, const Vec3& p2, const Vec3& p3)
@@ -41,97 +28,6 @@ scalar tetVolume(const Vec3& p0, const Vec3& p1, const Vec3& p2, const Vec3& p3)
 }
 
 } // anonymous namespace
-
-
-Vector<Vec3> computeFaceCentres(
-    const Executor& exec, const Vector<Vec3>& points, SegmentedVector<localIdx, localIdx>& faceNodes
-)
-{
-    localIdx nFaces = faceNodes.numSegments();
-    Vector<Vec3> faceCentres(exec, nFaces, Vec3 {0.0, 0.0, 0.0});
-
-    auto fcView = faceCentres.view();
-    auto ptsView = points.view();
-    auto fnView = faceNodes.view();
-
-    parallelFor(
-        exec,
-        {0, nFaces},
-        NEON_LAMBDA(const localIdx f) {
-            auto [start, end] = fnView.bounds(f);
-            Vec3 centre {0.0, 0.0, 0.0};
-            localIdx nNodes = end - start;
-            for (localIdx n = start; n < end; ++n)
-            {
-                centre = centre + ptsView[fnView.values[n]];
-            }
-            fcView[f] = centre * (1.0 / static_cast<scalar>(nNodes));
-        },
-        "computeFaceCentres"
-    );
-
-    return faceCentres;
-}
-
-
-Vector<Vec3> computeFaceAreas(
-    const Executor& exec,
-    const Vector<Vec3>& points,
-    SegmentedVector<localIdx, localIdx>& faceNodes,
-    const Vector<Vec3>& faceCentres
-)
-{
-    localIdx nFaces = faceNodes.numSegments();
-    Vector<Vec3> faceAreas(exec, nFaces, Vec3 {0.0, 0.0, 0.0});
-
-    auto faView = faceAreas.view();
-    auto ptsView = points.view();
-    auto fnView = faceNodes.view();
-    auto fcView = faceCentres.view();
-
-    parallelFor(
-        exec,
-        {0, nFaces},
-        NEON_LAMBDA(const localIdx f) {
-            auto [start, end] = fnView.bounds(f);
-            localIdx nNodes = end - start;
-            Vec3 area {0.0, 0.0, 0.0};
-            Vec3 centre = fcView[f];
-            for (localIdx n = 0; n < nNodes; ++n)
-            {
-                localIdx curr = start + n;
-                localIdx next = start + ((n + 1) % nNodes);
-                area = area
-                     + triangleArea(
-                           centre, ptsView[fnView.values[curr]], ptsView[fnView.values[next]]
-                     );
-            }
-            faView[f] = area;
-        },
-        "computeFaceAreas"
-    );
-
-    return faceAreas;
-}
-
-
-Vector<scalar> computeMagFaceAreas(const Executor& exec, const Vector<Vec3>& faceAreas)
-{
-    localIdx nFaces = faceAreas.size();
-    Vector<scalar> magFaceAreas(exec, nFaces);
-
-    auto magView = magFaceAreas.view();
-    auto faView = faceAreas.view();
-
-    parallelFor(
-        exec,
-        {0, nFaces},
-        NEON_LAMBDA(const localIdx f) { magView[f] = mag(faView[f]); },
-        "computeMagFaceAreas"
-    );
-
-    return magFaceAreas;
-}
 
 
 SegmentedVector<localIdx, localIdx> buildCellToFaceMapping(
@@ -282,59 +178,6 @@ Vector<scalar> computeCellVolumes(
     );
 
     return cellVolumes;
-}
-
-
-MeshGeometry computeGeometry(
-    const Executor& exec,
-    const Vector<Vec3>& points,
-    const Vector<localIdx>& faceOwner,
-    const Vector<localIdx>& faceNeighbour,
-    SegmentedVector<localIdx, localIdx>& faceNodes,
-    localIdx nInternalFaces,
-    localIdx nCells
-)
-{
-    auto faceCentres = computeFaceCentres(exec, points, faceNodes);
-    auto faceAreas = computeFaceAreas(exec, points, faceNodes, faceCentres);
-    auto magFaceAreasVec = computeMagFaceAreas(exec, faceAreas);
-
-    auto cellFaces = buildCellToFaceMapping(exec, faceOwner, faceNeighbour, nInternalFaces, nCells);
-    auto cellCentres = computeCellCentres(exec, faceCentres, cellFaces, nCells);
-    auto cellVolumes =
-        computeCellVolumes(exec, points, faceNodes, faceCentres, cellCentres, cellFaces, nCells);
-
-    return MeshGeometry {
-        std::move(cellVolumes),
-        std::move(cellCentres),
-        std::move(faceAreas),
-        std::move(faceCentres),
-        std::move(magFaceAreasVec)
-    };
-}
-
-
-MeshGeometry
-computeGeometry(const std::vector<Vec3>& points, const FaceTopology& topo, localIdx nCells)
-{
-    SerialExecutor exec;
-
-    // Convert points to NeoN Vector
-    Vector<Vec3> devicePoints(exec, points);
-
-    // faceOwner and faceNeighbour are already NeoN types in the new FaceTopology.
-    // faceNodes needs a non-const copy since computeGeometry takes SegmentedVector&.
-    auto faceNodesCopy = topo.faceNodes;
-
-    return computeGeometry(
-        exec,
-        devicePoints,
-        topo.faceOwner,
-        topo.faceNeighbour,
-        faceNodesCopy,
-        topo.nInternalFaces,
-        nCells
-    );
 }
